pyproject.toml
[project]
name = "quickbooks-mcp"
version = "1.0.0"
description = "MCP server for QuickBooks Online — 6 tools, ~120 operations across 20+ entity types + reports"
requires-python = ">=3.11"
dependencies = [
    "mcp>=1.0.0",
    "python-quickbooks>=0.9.12",
    "intuit-oauth>=1.2",
    "python-dotenv>=1.0.0",
]

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["server.py", "client.py", "auth_flow.py"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
